Use the custom choice for genome build
