+++ gens/commands/load.py
@@ -10,6 +10,7 @@ from flask import current_app as app
 from flask.cli import with_appcontext
 from pymongo.database import Database
 
+from gens.commands.util import ChoiceType
 from gens.config import settings
 from gens.db import (
     ANNOTATIONS_COLLECTION,
@@ -34,22 +35,6 @@ from gens.models.annotation import AnnotationRecord
 from gens.models.genomic import GenomeBuild
 
 LOG = logging.getLogger(__name__)
-
-
-class ChoiceType(click.Choice):
-    """Custom input type for click that returns genome build enum."""
-
-    name = "genome build"
-
-    def __init__(self, enum):
-        super().__init__(list(map(str, enum)))
-        self.enum = enum
-
-    def convert(self, value: str, param, ctx):
-        """Convert str to genome build"""
-
-        value = super().convert(value, param, ctx)
-        return next(v for v in self.enum if str(v) == value)
 
 
 def open_text_or_gzip(file_path: str) -> TextIO:
